requirements.txt
fastapi
uvicorn
openai
razorpay
python-dotenv
jinja2
aiofiles
httpx